+++ src/main/java/com/ggiiig/sse/SseRestController.java
@@ -1,7 +1,5 @@
 package com.ggiiig.sse;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ggiiig.webhook.dto.response.ResultDto;
 import groovy.util.logging.Slf4j;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,44 +8,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import java.io.IOException;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
 @lombok.extern.slf4j.Slf4j
 @Slf4j
 @RestController
 @RequestMapping("/sse")
 @RequiredArgsConstructor
 public class SseRestController {
-    private final ObjectMapper objectMapper;
-    // 연결 관리용 맵
-    private final Map<String, SseEmitter> emitters = new ConcurrentHashMap<>();
+    private final SseService sseService;
 
     @GetMapping("/result/{taskId}")
     public SseEmitter subscribe(@PathVariable String taskId) {
-        SseEmitter emitter = new SseEmitter(60 * 10000L);
-
-        emitters.put(taskId, emitter);
-
-        emitter.onCompletion(() -> emitters.remove(taskId));
-        emitter.onTimeout(() -> emitters.remove(taskId));
-        emitter.onError(e -> emitters.remove(taskId));
-
-        return emitter;
-    }
-
-    public void sendResult(String taskId, ResultDto result) {
-        SseEmitter emitter = emitters.get(taskId);
-        if (emitter != null) {
-            try {
-                emitter.send(SseEmitter.event().data(objectMapper.writeValueAsString(result)));
-                emitter.complete();
-            } catch (IOException e) {
-                log.error(e.getMessage(), e);
-            } finally {
-                emitters.remove(taskId);
-            }
-        }
+        return sseService.subscribe(taskId);
     }
 }
